Move exporter goods forms tests
